python/pyproject.toml
[project]
name = "simpleflow-sdk"
version = "0.1.0"
description = "SimpleFlow remote runtime SDK for Python"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0,<1.0.0",
  "PyJWT>=2.10.0,<3.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["simpleflow_sdk*"]
